move epoch_usec() to test.c
