+++ testprof/x86_prof.c
@@ -1,35 +1,5 @@
 #include <tomcrypt_test.h>
 
-#if defined(_WIN32)
-  #include <windows.h> /* GetSystemTimeAsFileTime */
-#else
-  #include <sys/time.h>
-#endif
-
-/* microseconds since 1970 (UNIX epoch) */
-ulong64 epoch_usec(void)
-{
-#if defined(LTC_NO_TEST_TIMING)
-  return 0;
-#elif defined(_WIN32)
-  FILETIME CurrentTime;
-  ulong64 cur_time;
-  ULARGE_INTEGER ul;
-  GetSystemTimeAsFileTime(&CurrentTime);
-  ul.LowPart  = CurrentTime.dwLowDateTime;
-  ul.HighPart = CurrentTime.dwHighDateTime;
-  cur_time = ul.QuadPart;
-  cur_time -= CONST64(116444736000000000); /* subtract epoch in microseconds */
-  cur_time /= 10; /* nanoseconds > microseconds */
-  return cur_time;
-#else
-  struct timeval tv;
-  struct timezone tz;
-  gettimeofday(&tv, &tz);
-  return (ulong64)(tv.tv_sec) * 1000000 + (ulong64)(tv.tv_usec); /* get microseconds */
-#endif
-}
-
 struct list results[100];
 int no_results;
 int sorter(const void *a, const void *b)
